Refactored deletion confirmation dialog for new state machine
