+++ app/src/main/java/ru/whoame/recogniser/AppModule.kt
@@ -6,10 +6,10 @@ import ru.whoame.recogniser.data.database.databaseModule
 import ru.whoame.recogniser.data.datasource.datasourceModule
 import ru.whoame.recogniser.data.repository.repositoryModule
 import ru.whoame.recogniser.ui.dialog.DeleteRecognisedItemViewModel
-import ru.whoame.recogniser.ui.dialog.factory.DeleteRecognisedItemDefaultsFactory
-import ru.whoame.recogniser.ui.dialog.factory.DeleteRecognisedItemStateFactory
-import ru.whoame.recogniser.ui.dialog.handler.DeleteRecognisedItemDomainEventHandler
-import ru.whoame.recogniser.ui.dialog.handler.DeleteRecognisedItemUiEventHandler
+import ru.whoame.recogniser.ui.dialog.state.factory.DeleteRecognisedItemDefaultsFactory
+import ru.whoame.recogniser.ui.dialog.state.factory.DeleteRecognisedItemStateFactory
+import ru.whoame.recogniser.ui.dialog.state.handler.DeleteRecognisedItemDomainEventHandler
+import ru.whoame.recogniser.ui.dialog.state.handler.DeleteRecognisedItemUiEventHandler
 import ru.whoame.recogniser.ui.screen.recognisedlist.RecognisedListViewModel
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.factory.ListScreenDefaultsFactory
 import ru.whoame.recogniser.ui.screen.recognisedlist.state.factory.ListScreenStateFactory
