--- app/src/main/java/ru/whoame/recogniser/ui/dialog/model/DeleteRecognisedItemDomainEvent.kt
@@ -1,7 +1,0 @@
-package ru.whoame.recogniser.ui.dialog.model
-
-sealed interface DeleteRecognisedItemDomainEvent {
-
-    data class Delete(val id: Long) : DeleteRecognisedItemDomainEvent
-
-}
