--- app/src/main/java/ru/whoame/recogniser/ui/dialog/model/DeleteRecognisedItemUiEvent.kt
@@ -1,7 +1,0 @@
-package ru.whoame.recogniser.ui.dialog.model
-
-sealed interface DeleteRecognisedItemUiEvent {
-
-    data class Delete(val id: Long) : DeleteRecognisedItemUiEvent
-
-}
