+++ app/src/main/java/ru/whoame/recogniser/ui/dialog/state/handler/DeleteRecognisedItemUiEventHandler.kt
@@ -1,9 +1,9 @@
-package ru.whoame.recogniser.ui.dialog.handler
+package ru.whoame.recogniser.ui.dialog.state.handler
 
 import kotlinx.coroutines.flow.Flow
-import ru.whoame.recogniser.ui.dialog.model.DeleteRecognisedItemDomainEvent
-import ru.whoame.recogniser.ui.dialog.model.DeleteRecognisedItemState
-import ru.whoame.recogniser.ui.dialog.model.DeleteRecognisedItemUiEvent
+import ru.whoame.recogniser.ui.dialog.state.model.DeleteRecognisedItemDomainEvent
+import ru.whoame.recogniser.ui.dialog.state.model.DeleteRecognisedItemState
+import ru.whoame.recogniser.ui.dialog.state.model.DeleteRecognisedItemUiEvent
 import ru.whoame.recogniser.utils.flowOf
 import ru.whoame.state_machine.handler.BaseUiEventHandler
 
@@ -13,7 +13,7 @@ class DeleteRecognisedItemUiEventHandler :
     override fun handleEvent(event: DeleteRecognisedItemUiEvent): Flow<*> = when (event) {
         is DeleteRecognisedItemUiEvent.Delete -> flowOf {
             reduceEvent {
-                DeleteRecognisedItemDomainEvent.Delete(event.id)
+                DeleteRecognisedItemDomainEvent.Delete
             }
         }
     }
