+++ app/src/main/java/ru/whoame/recogniser/ui/dialog/state/model/DeleteRecognisedItemDomainEvent.kt
@@ -1,0 +1,7 @@
+package ru.whoame.recogniser.ui.dialog.state.model
+
+sealed interface DeleteRecognisedItemDomainEvent {
+
+    data object Delete : DeleteRecognisedItemDomainEvent
+
+}
