+++ app/src/main/java/ru/whoame/recogniser/ui/dialog/state/model/DeleteRecognisedItemUiEvent.kt
@@ -1,0 +1,7 @@
+package ru.whoame.recogniser.ui.dialog.state.model
+
+sealed interface DeleteRecognisedItemUiEvent {
+
+    data object Delete : DeleteRecognisedItemUiEvent
+
+}
