+++ app/src/main/java/ru/whoame/recogniser/ui/dialog/state/model/DeleteRecognisedItemUiState.kt
@@ -1,10 +1,9 @@
-package ru.whoame.recogniser.ui.dialog.model
+package ru.whoame.recogniser.ui.dialog.state.model
 
 import androidx.compose.runtime.Immutable
 
 @Immutable
 data class DeleteRecognisedItemUiState(
     val itemName: String,
-    val itemId: Long,
     val isLoading: Boolean,
 )
